requirements.txt
argparse==1.4.0
tensorflow-gpu==2.3.0